Refactor: PlanetSelectorTopNavigation -> PlanetSelector
