+++ src/modules/PlanetSelector/index.ts
@@ -1,0 +1,1 @@
+export * from './PlanetSelector.client';
