--- src/modules/PlanetSelectorTopNavigation/index.ts
@@ -1,2 +1,0 @@
-export * from './PlanetSelector.client';
-export * from './PlanetSelectorTopNavigation';
